requirements.txt
requests-html
beautifulsoup4
nomic
cohere
dotenv
googlesearch-python
numpy
json